requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
python-dotenv


